# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Migration Factory API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/projects/{project_id}/effort": {
   "post": {
    "tags": [
     "effort"
    ],
    "summary": "Log Effort",
    "operationId": "log_effort_api_v1_projects__project_id__effort_post",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EffortIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EffortOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "effort"
    ],
    "summary": "List Effort",
    "operationId": "list_effort_api_v1_projects__project_id__effort_get",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/scans/{scan_id}/effort-estimate": {
   "get": {
    "tags": [
     "effort"
    ],
    "summary": "Scan Estimate",
    "operationId": "scan_estimate_api_v1_scans__scan_id__effort_estimate_get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/executions/{execution_id}/changes": {
   "get": {
    "tags": [
     "modernization"
    ],
    "summary": "Changes",
    "description": "Per-step change record: rule, file, lines, before/after (redacted), confidence level, snapshot and rollback patch.",
    "operationId": "changes_api_v1_executions__execution_id__changes_get",
    "parameters": [
     {
      "name": "execution_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Execution Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Changes Api V1 Executions  Execution Id  Changes Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/executions/{execution_id}/modernizations": {
   "post": {
    "tags": [
     "modernization"
    ],
    "summary": "Start",
    "operationId": "start_api_v1_executions__execution_id__modernizations_post",
    "parameters": [
     {
      "name": "execution_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Execution Id"
      }
     },
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ModernizationIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModernizationAccepted"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "modernization"
    ],
    "summary": "List For Execution",
    "operationId": "list_for_execution_api_v1_executions__execution_id__modernizations_get",
    "parameters": [
     {
      "name": "execution_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Execution Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ModernizationOut"
         },
         "title": "Response List For Execution Api V1 Executions  Execution Id  Modernizations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/modernizations/{modernization_id}": {
   "get": {
    "tags": [
     "modernization"
    ],
    "summary": "Get One",
    "operationId": "get_one_api_v1_modernizations__modernization_id__get",
    "parameters": [
     {
      "name": "modernization_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Modernization Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModernizationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EffortIn": {
    "properties": {
     "category": {
      "type": "string",
      "pattern": "^[a-z0-9:\\-/_A-Z]{2,60}$",
      "title": "Category"
     },
     "minutes": {
      "type": "integer",
      "maximum": 14400.0,
      "minimum": 1.0,
      "title": "Minutes"
     },
     "units": {
      "type": "integer",
      "maximum": 1000.0,
      "minimum": 1.0,
      "title": "Units",
      "default": 1
     },
     "subjectType": {
      "type": "string",
      "enum": [
       "finding",
       "diff_file",
       "integration",
       "other"
      ],
      "title": "Subjecttype",
      "default": "other"
     },
     "subjectId": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Subjectid"
     },
     "note": {
      "type": "string",
      "maxLength": 2000,
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "category",
     "minutes"
    ],
    "title": "EffortIn"
   },
   "EffortOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "category": {
      "type": "string",
      "title": "Category"
     },
     "minutes": {
      "type": "integer",
      "title": "Minutes"
     },
     "units": {
      "type": "integer",
      "title": "Units"
     },
     "subjectType": {
      "type": "string",
      "title": "Subjecttype"
     },
     "subjectId": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Subjectid"
     },
     "note": {
      "type": "string",
      "title": "Note"
     },
     "actor": {
      "type": "string",
      "title": "Actor"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "category",
     "minutes",
     "units",
     "subjectType",
     "subjectId",
     "note",
     "actor",
     "createdAt"
    ],
    "title": "EffortOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "JobOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "phase": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase"
     },
     "progress": {
      "type": "integer",
      "title": "Progress"
     },
     "attempts": {
      "type": "integer",
      "title": "Attempts"
     },
     "cancelRequested": {
      "type": "boolean",
      "title": "Cancelrequested"
     },
     "errorCode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errorcode"
     },
     "errorMessage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errormessage"
     },
     "createdBy": {
      "type": "string",
      "title": "Createdby"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "startedAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Startedat"
     },
     "finishedAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finishedat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind",
     "state",
     "phase",
     "progress",
     "attempts",
     "cancelRequested",
     "errorCode",
     "errorMessage",
     "createdBy",
     "createdAt",
     "startedAt",
     "finishedAt"
    ],
    "title": "JobOut"
   },
...
   "ModernizationAccepted": {
    "properties": {
     "modernizationId": {
      "type": "string",
      "format": "uuid",
      "title": "Modernizationid"
     },
     "jobId": {
      "type": "string",
      "format": "uuid",
      "title": "Jobid"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "modernizationId",
     "jobId",
     "status"
    ],
    "title": "ModernizationAccepted"
   },
   "ModernizationIn": {
    "properties": {
     "approve": {
      "items": {
       "type": "string",
       "pattern": "^MOD_[A-Z_]{2,40}$"
      },
      "type": "array",
      "maxItems": 20,
      "title": "Approve"
     }
    },
    "type": "object",
    "title": "ModernizationIn"
   },
   "ModernizationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "projectId": {
      "type": "string",
      "format": "uuid",
      "title": "Projectid"
     },
     "executionId": {
      "type": "string",
      "format": "uuid",
      "title": "Executionid"
     },
     "jobId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Jobid"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "approved": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Approved"
     },
     "results": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Results"
     },
     "qualityBefore": {
      "additionalProperties": true,
      "type": "object",
      "title": "Qualitybefore"
     },
     "qualityAfter": {
      "additionalProperties": true,
      "type": "object",
      "title": "Qualityafter"
     },
     "summary": {
      "additionalProperties": true,
      "type": "object",
      "title": "Summary"
     },
     "createdBy": {
      "type": "string",
      "title": "Createdby"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "job": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "projectId",
     "executionId",
     "jobId",
     "state",
     "approved",
     "results",
     "qualityBefore",
     "qualityAfter",
     "summary",
     "createdBy",
     "createdAt"
    ],
    "title": "ModernizationOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}